Move FetchMetrics to hdds.scm package
